+++ adaptivetesting/data/__sqlite_context.py
@@ -1,7 +1,7 @@
 from typing import List
-from ..models.adaptive_test import AdaptiveTest
-from ..models.test_result import TestResult
-from ..services.test_results_interface import ITestResults
+from ..models.__adaptive_test import AdaptiveTest
+from ..models.__test_result import TestResult
+from ..services.__test_results_interface import ITestResults
 import sqlite3
