+++ adaptivetesting/implementations/__pre_test.py
@@ -1,6 +1,6 @@
 from typing import List
 import numpy as np
-from ..models.test_item import TestItem
+from ..models.__test_item import TestItem
 import random
