+++ adaptivetesting/math/__urrys_rule.py
@@ -1,4 +1,4 @@
-from ..models.test_item import TestItem
+from ..models.__test_item import TestItem
 from typing import List
