+++ adaptivetesting/models/__init__.py
@@ -1,5 +1,5 @@
-from .adaptive_test import *
-from .algorithm_exception import *
-from .item_pool import *
-from .test_item import *
-from .test_result import *
+from .__adaptive_test import *
+from .__algorithm_exception import *
+from .__item_pool import *
+from .__test_item import *
+from .__test_result import *
