+++ adaptivetesting/models/__item_pool.py
@@ -1,4 +1,4 @@
-from .test_item import TestItem
+from .__test_item import TestItem
 from typing import List, overload, Tuple
