+++ adaptivetesting/simulation/__init__.py
@@ -1,1 +1,1 @@
-from .simulation import *
+from .__simulation import *
